fix: clean examination session page
