+++ src/features/exam-sessions/index.ts
@@ -1,0 +1,1 @@
+export { ExamSessionsPage } from './ui/ExamSessionsPage'
